requirements.txt
matplotlib==3.7.2
peewee==3.16.2
pyTelegramBotAPI==4.12.0
python-dotenv==1.0.0
python-telegram-bot-calendar==1.0.5